SplashContent: remove event listeners on unmount
The "dragover" and "drop" event listeners are only needed on the splash
page where users drag-and-drop their dataset. Once the data is loaded,
the `SplashContent` component is unmounted and the event listeners are
explicitly removed as well.

This will stop the unexpected error notification for dragged on metadata
files that triggered both the auspice.us and auspice event listeners.