Store datasets as auspice Dataset structures
Using `Dataset` objects will pave the way for sidecar files as well as caching datasets in auspice which is necessary for narrative functionality. 

No behaviour is expected to change in this commit beyond notification messages & logging. 

Note that this requires an auspice version of 2.34.1 or later which exports the `Dataset` object.