Handle dropped narrative
Narratives are complicated as slides are linked to a URL not a filepath. This implementation requires that a dataset URL "a/b/c" (for a given narrative slide) is matched to a dropped file "a_b_c.json".

There are some minor papercuts experienced while testing this, such as URL queries being added, and auspice partially freezing when a dataset in the narrative is not present in the dragged files. Fixing these requires work in auspice & for the time being I don't think this is warranted.