Improve Newick parsing
Replaces the basic Newick parser with an external one that handles
single-quoted node names.  Quoted names are allowed by the format¹ and
occur in trees produced by NCBI Pathogens.²

The parser's API is a little awkward for our use case, but it's
perfectly workable.  Out of several parsers I tried on NPM, this was the
only one which handled quoted names, so use it despite the slightly
awkward API.

¹ See <https://en.wikipedia.org/wiki/Newick_format#Notes> for lack of
  any formal spec.

² <https://discussion.nextstrain.org/t/displaying-trees-from-ncbi-pathogen-browser-in-auspice-us/1456>