Disallow newick trees with quotes
The newick parser we use is known to have issues with quoted node names,
as it will parse the node name content as newick. As a quick fix,
disallow trees with quotes. The result is that the trees in #66 and
<https://discussion.nextstrain.org/t/displaying-trees-from-ncbi-pathogen-browser-in-auspice-us/1456/4>
will now result in an error. The downside is trees with quoted names but
no newick-like structure within the node name will also fail to load.