Refactor processing of dropped files
Motivated by the desire to expand our allowable file suffixes (#126)
this simplifies what was a convoluted piece of code. Error messages
have changed slightly as part of this work, but the intention is for
there to be no other changes.

As part of testing this refactoring I found that narratives weren't
currently working - #128 - and that's fixed here.