Rearrange dropped file code
As suggested in review <https://github.com/nextstrain/auspice.us/pull/129#discussion_r2153067497>
it's nicer to co-locate the file reading code inside the function which
examines the dropped file. Note that we could turn this function into a
constructor, call it `DroppedFile` and invoke it with `new`; with the
code as it is currently architectured there is no difference.